Need to implement reccursive recognition.
