+++ headers/Config.hpp
@@ -4,22 +4,22 @@
 #include "Webserv.hpp"
 #include <fstream>
 #include <sstream>
-
+#include <deque>
 
 typedef struct s_server{
-	std::string	_port;
-	std::string	_location;
-	std::string	_server_name;
-	std::string	_root;
-	std::string	_include;
+	std::string					_port;
+	std::vector<std::string>	_location;
+	std::string					_server_name;
+	std::string					_root;
+	std::string					_include;
 }	server;
 
 class Config {
 private:
 	std::vector<server>	_Servers;
+	Config();
 
 public:
-	Config();
 	Config(const char* configFile);
 	~Config();
 	Config(const Config& c);
@@ -27,7 +27,8 @@ public:
 
 	// Parser main functions
 	void getConfig(std::string& confString);
-	void getServerConfig(const std::string& filebuff);
+	void getServerConfig(std::string& filebuff);
+	std::& getBlocks(std::string& confString);
 	
 	// Parser helper funcitons
